replace testutils with mocks
