+++ bctl/daemon/plugin/kube/actions/portforward/portforward_test.go
@@ -11,11 +11,11 @@ import (
 	"testing"
 	"time"
 
+	"bastionzero.com/bctl/v1/bzerolib/mocks"
 	"bastionzero.com/bctl/v1/bzerolib/plugin"
 	"bastionzero.com/bctl/v1/bzerolib/plugin/kube/actions/portforward"
 	kubeutils "bastionzero.com/bctl/v1/bzerolib/plugin/kube/utils"
 	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
-	"bastionzero.com/bctl/v1/bzerolib/testutils"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/build/kubernetes/api"
 	"gopkg.in/tomb.v2"
@@ -28,7 +28,7 @@ func setPerformHandshake() {
 	}
 }
 
-func setGetUpgradedConnection(mockConnection *testutils.MockStreamConnection) (http.Header, http.Header) {
+func setGetUpgradedConnection(mockConnection *mocks.MockStreamConnection) (http.Header, http.Header) {
 	dataHeaders := http.Header{}
 	dataHeaders.Set(kubeutils.StreamType, kubeutils.StreamTypeData)
 	errorHeaders := http.Header{}
@@ -65,7 +65,7 @@ func TestMain(m *testing.M) {
 func TestPortForward(t *testing.T) {
 	assert := assert.New(t)
 	var tmb tomb.Tomb
-	logger := testutils.MockLogger()
+	logger := mocks.MockLogger()
 	requestId := "rid"
 	logId := "lid"
 	command := "logs"
@@ -78,26 +78,26 @@ func TestPortForward(t *testing.T) {
 	headers := http.Header{}
 	headers.Set(kubeutils.PortForwardRequestIDHeader, requestId)
 
-	request := testutils.MockHttpRequest("GET", urlPath, headers, sendData)
+	request := mocks.MockHttpRequest("GET", urlPath, headers, sendData)
 
-	mockStreamConnection := testutils.MockStreamConnection{}
+	mockStreamConnection := mocks.MockStreamConnection{}
 	mockStreamConnection.On("SetIdleTimeout", kubeutils.DefaultIdleTimeout).Return()
 	var closeChan <-chan bool
 	mockStreamConnection.On("CloseChan").Return(closeChan)
 	mockStreamConnection.On("Close").Return(nil)
 
-	writer := testutils.MockResponseWriter{}
+	writer := mocks.MockResponseWriter{}
 
 	setPerformHandshake()
 	dataHeaders, errorHeaders := setGetUpgradedConnection(&mockStreamConnection)
 
-	mockDataStream := testutils.MockStream{MyStreamData: streamData}
+	mockDataStream := mocks.MockStream{MyStreamData: streamData}
 	mockDataStream.On("Headers").Return(dataHeaders)
 	mockDataStream.On("Close").Return(nil)
 	mockDataStream.On("Read", make([]byte, portforward.DataStreamBufferSize)).Return(len(streamData), nil)
 	mockDataStream.On("Write", []byte(testData)).Return(len(testData), nil).Times(3)
 
-	mockErrorStream := testutils.MockStream{MyStreamData: streamError}
+	mockErrorStream := mocks.MockStream{MyStreamData: streamError}
 	mockErrorStream.On("Headers").Return(errorHeaders)
 	mockErrorStream.On("Close").Return(nil)
 	mockErrorStream.On("Read", make([]byte, portforward.ErrorStreamBufferSize)).Return(len(streamError), nil)
@@ -213,7 +213,7 @@ func TestPortForward(t *testing.T) {
 func TestPortForwardError(t *testing.T) {
 	assert := assert.New(t)
 	var tmb tomb.Tomb
-	logger := testutils.MockLogger()
+	logger := mocks.MockLogger()
 	requestId := "rid"
 	logId := "lid"
 	command := "logs"
@@ -223,9 +223,9 @@ func TestPortForwardError(t *testing.T) {
 	// TODO: check validity
 	p, outputChan := New(logger, requestId, logId, command)
 
-	request := testutils.MockHttpRequest("GET", urlPath, make(map[string][]string), sendData)
+	request := mocks.MockHttpRequest("GET", urlPath, make(map[string][]string), sendData)
 
-	writer := testutils.MockResponseWriter{}
+	writer := mocks.MockResponseWriter{}
 	writer.On("WriteHeader", http.StatusForbidden).Return()
 	writer.On("Header").Return(make(map[string][]string))
 	// not crazy about this since we have to reuse code from the file we're testing
