+++ bzerolib/mocks/mocks.go
@@ -1,13 +1,28 @@
-package testutils
+package mocks
 
 import (
+	"bytes"
+	"encoding/base64"
 	"io"
+	"io/ioutil"
 	"net/http"
+	"net/url"
 	"time"
 
+	"bastionzero.com/bctl/v1/bzerolib/logger"
+	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"k8s.io/apimachinery/pkg/util/httpstream"
 )
+
+func MockLogger() *logger.Logger {
+	logger, err := logger.New(logger.DefaultLoggerConfig(logger.Debug.String()), "/dev/null")
+	if err == nil {
+		return logger
+	}
+	return nil
+}
 
 // FIXME: docstring
 type MockResponseWriter struct {
@@ -86,4 +101,24 @@ func (m *MockStreamConnection) CloseChan() <-chan bool {
 }
 func (m *MockStreamConnection) SetIdleTimeout(timeout time.Duration) {
 	m.Called(timeout)
+}
+
+// create an http request with the specified details
+func MockHttpRequest(method string, path string, headers map[string][]string, content string) http.Request {
+	return http.Request{
+		Method:        method,
+		URL:           &url.URL{Path: path},
+		Header:        headers,
+		ContentLength: int64(len(content)),
+		Body:          ioutil.NopCloser(bytes.NewBufferString(content)),
+	}
+}
+
+// assert that the content of the stream message coming from outputChan is equal to testSTring
+// this is a common pattern when testing the agent
+func AssertNextMessageHasContent(assert *assert.Assertions, outputChan chan smsg.StreamMessage, testString string) {
+	message := <-outputChan
+	content, err := base64.StdEncoding.DecodeString(message.Content)
+	assert.Nil(err)
+	assert.Equal([]byte(testString), content)
 }
